services/schemas.py
from pydantic import BaseModel
from typing import Any


class ChatMessage(BaseModel):
    role: str
    content: str


class ChatRequest(BaseModel):
    question: str
    history: list[ChatMessage] = []
